// src/app/models/etudiant.model.ts
export class Etudiant {
    id?: number = 0;
    firstname: string = '';
    lastname: string = '';
    firstnameA: string = '';
    lastnameA: string = '';
    birthdate: string = '';
    gender: string = '';
    cin: string = '';
    niveau: string = '';
    parcours : string='';
    moyenne: number = 0;
    profile: string = '';
  }